Adds authorization and roles with pundit & rolify

+++ app/controllers/characters_controller.rb
@@ -19,24 +19,14 @@ class CharactersController < ApplicationController
 
   # GET /characters/1/edit
   def edit
-    @game = Game.find params[:game_id]
-    maintainers = @game.maintainers.ids
-
-    if maintainers.exclude?(current_operator.id)
-      return redirect_back fallback_location: root_path
-    end
+    @game = @character.game
   end
 
   # POST /characters
   # POST /characters.json
   def create
-    @game = Game.find character_params[:game_id]
-    maintainers_ids = @game.maintainers.ids
-
-    if maintainers_ids.include?(current_operator.id)
-      @character = Character.new(character_params)
-      @character.save
-    end
+    @character = Character.new(character_params)
+    add_contributor(@character.game) if @character.save
 
     redirect_back fallback_location: root_path
   end
@@ -44,22 +34,19 @@ class CharactersController < ApplicationController
   # PATCH/PUT /characters/1
   # PATCH/PUT /characters/1.json
   def update
-    @game = Game.find character_params[:game_id]
-    maintainers = @game.maintainers.ids
-
-    if maintainers.include?(current_operator.id)
-      if @character.update(character_params)
-        flash[:notice] = "Character was successfully created."
-      else 
-        flash[:notice] = "Character was successfully created."
-      end
-    end 
-    redirect_to game_path(@game)
+    authorize @character
+    if @character.update(character_params)
+      flash[:notice] = "Character was successfully created."
+    else 
+      flash[:notice] = "Character was successfully created."
+    end
+    redirect_to game_path(@character.game)
   end
 
   # DELETE /characters/1
   # DELETE /characters/1.json
   def destroy
+    authorize @character
     if @character.destroy
       flash[:notice] = "Character was successfully destroyed."
     else
@@ -76,6 +63,7 @@ class CharactersController < ApplicationController
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def character_params
-      params.require(:character).permit(:game_id, :name)
+      params[:character][:operator_id] = current_operator.id
+      params.require(:character).permit(:game_id, :name, :operator_id)
     end
 end
